Adjust Main and Testings

+++ LAB_03/mySAG.h
@@ -18,7 +18,7 @@ void MySAGInsert(int);
 
 int MySAGMax();
 int MySAGMin();
-double MySAGAvg();
+int MySAGAvg();
 
 int MySAGFreq(int);
